made modificiations necessary to reuse existing server, test in the works
