+++ node/node.go
@@ -18,6 +18,7 @@ type Node interface {
 	GetClient() rpcclient.Client
 	Stop() error
 	IsRunning() bool
+	ReceiveDirectTx() func([]byte) bool
 }
 
 func NewNode(
